+++ apps/web/src/app/partner/layout.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from "react";
-import { PartnerHeader } from "@/shared/ui/PartnerHeader";
+import { PartnerHeader } from "./_components/PartnerHeader";
 import { Footer } from "@/shared/ui/Footer";
 
 /** 파트너(손해사정사) 페이지 레이아웃 셸 — 파트너 헤더 + 본문 + 푸터. */
